+++ Ironwall.Dotnet.Libraries.Api/Services/IApiService.cs
@@ -6,6 +6,7 @@ public interface IApiService : IService
 {
     void Initialize();
     Task<HttpResponseMessage> DeleteRequestAsync(string endpoint);
+    Task<HttpResponseMessage> DeleteRequestAsync<T>(string endpoint, T body);
     Task<HttpResponseMessage> GetRequestAsync(string endpoint, Dictionary<string, string>? parameters = null);
     Task<HttpResponseMessage> PatchRequestAsync<T>(string endpoint, T body);
     Task<HttpResponseMessage> PostFormDataRequestAsync(string endpoint, MultipartFormDataContent content);